Update neonDIDupdate to support Neon pagination
we have a lot of Neon accounts now -- 200-item pages don't cut it anymore.
Also, add util.py to hold common boilerplate.  Right now it's just the apiCall fn.
