+++ examples/util.py
@@ -1,0 +1,22 @@
+import requests
+
+## Helper function for API calls
+def apiCall(httpVerb, url, data, headers):
+    # Make request
+    if httpVerb == 'GET':
+        response = requests.get(url, data=data, headers=headers)
+    elif httpVerb == 'POST':
+        response = requests.post(url, data=data, headers=headers)
+    elif httpVerb == 'PUT':
+        response = requests.put(url, data=data, headers=headers)
+    elif httpVerb == 'PATCH':
+        response = requests.patch(url, data=data, headers=headers)
+    elif httpVerb == 'DELETE':
+        response = requests.delete(url, data=data, headers=headers)
+    else:
+        print(f"HTTP verb {httpVerb} not recognized")
+
+    response = response.json()
+    #pprint(response)
+
+    return response
